fix linux tun test clippy

+++ crates/tun/src/linux.rs
@@ -789,7 +789,7 @@ mod tests {
             (libc::IFF_TUN | libc::IFF_NO_PI) as libc::c_short | 0x4000
         );
         assert_eq!(TUN_F_CSUM | TUN_F_TSO4 | TUN_F_TSO6, 0x07);
-        assert_eq!(libc::TUNSETOFFLOAD as u64, 0x4004_54d0);
+        assert_eq!(libc::TUNSETOFFLOAD, 0x4004_54d0);
         assert_eq!(VNET_READ_LEN, offload::VIRTIO_NET_HDR_LEN + 65_535);
     }
